Create PostTypeIcon component
